+++ fabric-game-rule-api-v1-bridge/src/main/java/net/fabricmc/fabric/api/gamerule/v1/rule/ValidateableRule.java
@@ -1,0 +1,2 @@
+package net.fabricmc.fabric.api.gamerule.v1.rule;
+public interface ValidateableRule { boolean validate(String value); }